+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 Newest updates are at the top of this file.
 
+## xxx xx 2021 - v5.xxx
+* Add a sample amqsbo.go to show how to deal with poison messages
+
 ## Mar 25 2021 - v5.2.0 
 Scope of mqmetric changes seem to justify new minor number
 * Update for MQ 9.2.2